web_discovery/requirements.txt
kafka-python
requests
pattern
pymongo